refactor(footer): re-org file-structure into DesktopView/MobileView
