+++ src/containers/unit/Footer/DesktopView/DigestView/ContactBar.js
@@ -3,8 +3,11 @@ import React from 'react'
 import { useMedia } from '@/hooks'
 import EmailSubscriber from '@/components/EmailSubscriber'
 
-import SocialList from '../SocialList'
-import { Wrapper, InnerWrapper } from '../styles/digest_view/contact_bar'
+import SocialList from '../../SocialList'
+import {
+  Wrapper,
+  InnerWrapper,
+} from '../../styles/desktop_view/digest_view/contact_bar'
 
 const ContactBar = ({ layout }) => {
   const { mobile } = useMedia()
