+++ src/containers/unit/Footer/index.js
@@ -4,92 +4,21 @@
  *
  */
 
-import React, { useState, useEffect } from 'react'
-import { useRouter } from 'next/router'
-import dynamic from 'next/dynamic'
-import { contains } from 'ramda'
+import React from 'react'
 
-import { ROUTE } from '@/constant'
-import { connectStore, buildLog, getRoutePathList } from '@/utils'
+import { useMedia } from '@/hooks'
 
-import JoinModal from '@/containers/tool/JoinModal'
-import Modal from '@/components/Modal'
+import DesktopView from './DesktopView'
+import MobileView from './MobileView'
 
-import BriefView from './BriefView'
-import DigestView from './DigestView'
-import BusinessNote from './BusinessNote'
-
-import { Wrapper } from './styles'
-import {
-  useInit,
-  toggleSponsorHelper,
-  toggleBusBanner,
-  onLogin,
-  onPay,
-} from './logic'
-
-/* eslint-disable-next-line */
-const log = buildLog('C:Footer')
-
-export const BuyMeChuanChuan = dynamic(
-  () => import('@/components/BuyMeChuanChuan'),
-  {
-    /* eslint-disable react/display-name */
-    loading: () => <div />,
-    ssr: false,
-  },
-)
-
-const FooterContainer = ({ footer: store }) => {
-  useInit(store)
-  const {
-    showSponsor,
-    showBusBanner,
-    accountInfo,
-    accountInfo: {
-      customization: { bannerLayout },
-    },
-  } = store
-
-  const router = useRouter()
-  const [mainPath, subPath] = getRoutePathList(router.asPath)
-
-  const [curView, setCurView] = useState('DIGEST')
-
-  useEffect(() => {
-    if (
-      contains(mainPath, [ROUTE.USER, ROUTE.DISCOVERY]) ||
-      contains(subPath, [ROUTE.POST, ROUTE.JOB, ROUTE.VIDEO, ROUTE.REPO])
-    ) {
-      setCurView('BRIEF')
-    } else if (contains(mainPath, [ROUTE.SUBSCRIBE])) {
-      setCurView('BRIEF_LEFT')
-    } else {
-      setCurView('DIGEST')
-    }
-  }, [mainPath, subPath])
+const FooterContainer = (props) => {
+  const { mobile } = useMedia()
 
   return (
-    <Wrapper testId="footer" layout={bannerLayout}>
-      <Modal show={showBusBanner} showCloseBtn onClose={toggleBusBanner}>
-        <BusinessNote />
-      </Modal>
-      <JoinModal />
-      <BuyMeChuanChuan
-        show={showSponsor}
-        accountInfo={accountInfo}
-        onClose={toggleSponsorHelper}
-        onLogin={onLogin}
-        onPay={onPay}
-      />
-
-      {curView === 'DIGEST' ? (
-        <DigestView layout={bannerLayout} />
-      ) : (
-        <BriefView curView={curView} />
-      )}
-    </Wrapper>
+    <React.Fragment>
+      {!mobile ? <DesktopView {...props} /> : <MobileView {...props} />}
+    </React.Fragment>
   )
 }
 
-export default connectStore(FooterContainer)
+export default FooterContainer
