+++ src/containers/unit/Footer/styles/desktop_view/bottom_info.js
