+++ src/containers/unit/Footer/styles/mobile_view.js
